requirements.txt
Django==1.7.4
uWSGI==2.0.9
Fabric==1.10.1
Pillow==2.7.0
psycopg2==2.5.4
django-tagging==0.3.4
csscompressor==0.9.3
